Factor out.
