+++ impl/ZOWIE.hs/src/Language/ZOWIE/Machine.hs
@@ -1,0 +1,7 @@
+module Language.ZOWIE.Machine where
+
+import Language.ZOWIE.Registers
+import Language.ZOWIE.State
+
+
+run s = s
